pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "bughunt-env"
version = "1.0.0"
description = "OpenEnv RL environment — find and fix bugs in Python code"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
